Refactor files and delete old functions and files
