+++ app.py
@@ -1,6 +1,6 @@
 import os
 
-# load environemnt variables from .env file
+# load environemnt variables from .env file (local build)
 if os.path.exists(".env"):
     with open(".env") as f:
         for line in f:
@@ -9,33 +9,27 @@ if os.path.exists(".env"):
                 os.environ.setdefault(key,value.strip().strip('"').strip("'"))
 
 import pymysql
-from flask import Flask, render_template, request, jsonify, session
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-from src.ball24 import TippeData24
-from src.ball25 import TippeData25
-from src.ball26 import TippeData26
+from flask import Flask, render_template
+from src.ball.ball24 import TippeData24
+from src.ball.ball25 import TippeData25
+from src.ball.ball26 import TippeData26
 from src.reader import CsvKampspill
 from src.db import init_db
-from src.routes import bets_bp, matches_bp, auth_bp, register_bp
+from src.routes import register_blueprints
 import src.app_globals
 pymysql.install_as_MySQLdb()
 
 app = Flask(__name__)
-
-
 
 # Load environment variables
 app.secret_key = os.getenv('FIFAGUTTA_SECRET_KEY')
 
 # set up database and register blueprints
 init_db(app)
-app.register_blueprint(auth_bp)
-app.register_blueprint(register_bp)
-app.register_blueprint(bets_bp)
-app.register_blueprint(matches_bp)
+register_blueprints(app)
 
 app.jinja_env.globals["g_PRESEASON"] = src.app_globals.PRESEASON
+
 
 # PRESEASON
 if src.app_globals.PRESEASON:
@@ -104,6 +98,7 @@ def r24():
 @app.route('/2023')
 def hjelp():
 	return "vi rykket ned"
+
 
 if __name__ == "__main__":
 	app.run(host="0.0.0.0", debug=True)
