+++ src/ball/ball26.py
@@ -1,12 +1,11 @@
 import data.tips26
-from .tippedata import TippeData
-from .common import Team
+import tippedata
+from ..common import Team
 import os
-import csv
 import datetime
 import shutil
 
-class TippeData26(TippeData):
+class TippeData26(tippedata.TippeData):
     def __init__(self, debug=False):
         year = 2026
         entries = data.tips26.ENTRIES
