--- templates/index.html
@@ -1,93 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>Fifagutta</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-  <link rel="shortcut icon" type="image/x-icon" href="/static/olav.ico">
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.9.4/Chart.js"></script>
-  <!-- Optionally include any other shared CSS -->
-  <style>
-    body,h1,h2,h3,h4,h5,h6 {font-family: "Lato", sans-serif;}
-    body, html {
-      height: 100%;
-      color: #00538a;
-      line-height: 1.5;
-    }
-    /* Common styles for tabs */
-    .tab {
-      overflow: hidden;
-      background-color: #f1f1f1;
-    }
-    .tab button {
-      background-color: inherit;
-      border: none;
-      outline: none;
-      cursor: pointer;
-      padding: 14px 16px;
-      transition: 0.3s;
-      font-size: 17px;
-    }
-    .tab button:hover {
-      background-color: #ddd;
-    }
-    .tab button.active {
-      background-color: #ccc;
-    }
-    .tabcontent {
-      display: none;
-      padding: 20px;
-    }
-    /* Navbar */
-    .w3-bar { display: flex; justify-content: center; background-color: #f0f0f0; }
-    .w3-bar .tablink { font-weight: bold; margin: 0 10px; }
-    /* For mobile: enable horizontal scrolling if needed */
-    @media (max-width: 1200px) {
-      .scrollable-table {
-        overflow-x: auto;
-      }
-    }
-  </style>
-  <script>
-    function openTab(evt, tabName) {
-      var i, tabcontent, tablinks;
-      tabcontent = document.getElementsByClassName("tabcontent");
-      for (i = 0; i < tabcontent.length; i++) {
-        tabcontent[i].style.display = "none";
-      }
-      tablinks = document.getElementsByClassName("tablink");
-      for (i = 0; i < tablinks.length; i++) {
-        tablinks[i].className = tablinks[i].className.replace(" active", "");
-      }
-      document.getElementById(tabName).style.display = "block";
-      evt.currentTarget.className += " active";
-    }
-    window.onload = function() {
-      // Open default tab: current year results (obos25.html)
-      document.getElementById("defaultOpen").click();
-    }
-  </script>
-</head>
-<body>
-  <!-- Common header/navigation -->
-  <div class="w3-top">
-    <div class="w3-bar" id="navBarFull">
-      <button class="tablink" id="defaultOpen" onclick="openTab(event, 'current')">Current Year</button>
-      <button class="tablink" onclick="openTab(event, 'last')">Last Year</button>
-    </div>
-  </div>
-
-  <!-- Tab content: current year -->
-  <div id="current" class="tabcontent">
-    {% include 'obos25.html' %}
-  </div>
-
-  <!-- Tab content: last year -->
-  <div id="last" class="tabcontent">
-    {% include 'r-24.html' %}
-  </div>
-</body>
-</html>
